Patch for 307580

+++ bundles/org.eclipse.equinox.p2.touchpoint.natives/src/org/eclipse/equinox/internal/p2/touchpoint/natives/Util.java
@@ -13,6 +13,7 @@ package org.eclipse.equinox.internal.p2.touchpoint.natives;
 import java.io.*;
 import java.net.URI;
 import java.util.*;
+import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 import org.eclipse.core.runtime.*;
@@ -81,9 +82,18 @@ public class Util {
 	 * monitor and backup store may be null.
 	 */
 	public static File[] unzipFile(File zipFile, File outputDir, IBackupStore store, String taskName, IProgressMonitor monitor) throws IOException {
+		return unzipFile(zipFile, outputDir, null, null, store, taskName, monitor);
+	}
+
+	/**
+	 * Unzip from a File to an output directory, with progress indication and backup.
+	 * It takes in count exclude/exclude pattern (that can be null, case when everything is unzipped)
+	 * monitor and backup store may be null.
+	 */
+	public static File[] unzipFile(File zipFile, File outputDir, String[] includePatterns, String[] excludePatterns, IBackupStore store, String taskName, IProgressMonitor monitor) throws IOException {
 		InputStream in = new FileInputStream(zipFile);
 		try {
-			return unzipStream(in, zipFile.length(), outputDir, store, taskName, monitor);
+			return unzipStream(in, zipFile.length(), outputDir, includePatterns, excludePatterns, store, taskName, monitor);
 		} catch (IOException e) {
 			// add the file name to the message
 			throw new IOException(NLS.bind(Messages.Util_Error_Unzipping, zipFile, e.getMessage()));
@@ -97,6 +107,14 @@ public class Util {
 	 * if backup store is not null.
 	 */
 	public static File[] unzipStream(InputStream stream, long size, File outputDir, IBackupStore store, String taskName, IProgressMonitor monitor) throws IOException {
+		return unzipStream(stream, size, outputDir, null, null, store, taskName, monitor);
+	}
+
+	/**
+	 * Unzip from an InputStream to an output directory using backup of overwritten files
+	 * if backup store is not null.
+	 */
+	public static File[] unzipStream(InputStream stream, long size, File outputDir, String[] includePatterns, String[] excludePatterns, IBackupStore store, String taskName, IProgressMonitor monitor) throws IOException {
 		InputStream is = monitor == null ? stream : stream; // new ProgressMonitorInputStream(stream, size, size, taskName, monitor); TODO Commented code
 		ZipInputStream in = new ZipInputStream(new BufferedInputStream(is));
 		ZipEntry ze = in.getNextEntry();
@@ -106,28 +124,61 @@ public class Util {
 			in.close();
 			throw new IOException(Messages.Util_Invalid_Zip_File_Format);
 		}
+		Collection<Pattern> includeRegexp = new ArrayList<Pattern>();
+		Collection<Pattern> excludeRegexp = new ArrayList<Pattern>();
+		if (includePatterns != null) {
+			for (String pattern : includePatterns) {
+				if (pattern != null) {
+					includeRegexp.add(createAntStylePattern(pattern));
+				}
+			}
+		}
+		if (excludePatterns != null) {
+			for (String pattern : excludePatterns) {
+				if (pattern != null) {
+					excludeRegexp.add(createAntStylePattern(pattern));
+				}
+			}
+		}
 		ArrayList<File> unzippedFiles = new ArrayList<File>();
 		do {
-			File outFile = new File(outputDir, ze.getName());
-			unzippedFiles.add(outFile);
-			if (ze.isDirectory()) {
-				outFile.mkdirs();
-			} else {
-				if (outFile.exists()) {
-					if (store != null)
-						store.backup(outFile);
-					else
-						outFile.delete();
+			String name = ze.getName();
+			boolean unzip = includeRegexp.isEmpty();
+			for (Pattern pattern : includeRegexp) {
+				unzip = pattern.matcher(name).matches();
+				if (unzip)
+					break;
+			}
+			if (unzip && !excludeRegexp.isEmpty()) {
+				for (Pattern pattern : excludeRegexp) {
+					if (pattern.matcher(name).matches()) {
+						unzip = false;
+						break;
+					}
+				}
+			}
+			if (unzip) {
+				File outFile = new File(outputDir, name);
+				unzippedFiles.add(outFile);
+				if (ze.isDirectory()) {
+					outFile.mkdirs();
 				} else {
-					outFile.getParentFile().mkdirs();
+					if (outFile.exists()) {
+						if (store != null)
+							store.backup(outFile);
+						else
+							outFile.delete();
+					} else {
+						outFile.getParentFile().mkdirs();
+					}
+					try {
+						copyStream(in, false, new FileOutputStream(outFile), true);
+					} catch (FileNotFoundException e) {
+						// TEMP: ignore this for now in case we're trying to replace
+						// a running eclipse.exe
+					}
+					outFile.setLastModified(ze.getTime());
 				}
-				try {
-					copyStream(in, false, new FileOutputStream(outFile), true);
-				} catch (FileNotFoundException e) {
-					// TEMP: ignore this for now in case we're trying to replace
-					// a running eclipse.exe
-				}
-				outFile.setLastModified(ze.getTime());
 			}
 			in.closeEntry();
 		} while ((ze = in.getNextEntry()) != null);
@@ -163,4 +214,31 @@ public class Util {
 			}
 		}
 	}
+
+	private static Pattern createAntStylePattern(String pattern) {
+		StringBuffer sb = new StringBuffer();
+		for (int c = 0; c < pattern.length(); c++) {
+			switch (pattern.charAt(c)) {
+				case '.' :
+					sb.append("\\."); //$NON-NLS-1$
+					break;
+				case '*' :
+					sb.append(".*"); //$NON-NLS-1$
+					break;
+				case '?' :
+					sb.append(".?"); //$NON-NLS-1$
+					break;
+				default :
+					sb.append(pattern.charAt(c));
+					break;
+			}
+		}
+		String string = sb.toString();
+		if (string.endsWith("\\..*")) { //$NON-NLS-1$
+			sb.append("|"); //$NON-NLS-1$
+			sb.append(string.substring(0, string.length() - 4));
+		}
+		return Pattern.compile(sb.toString());
+	}
+
 }
